+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to the "csharp-bootstrapper" extension will be documented in
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## 0.0.0 - 0.0.4
 
-- Initial release
+- Initial scaffolding of the project
+- Added Activity bar icon with welcome view
+- Added configuration (Frontend model directory)
+- C# Context menu and Editor menu command to generate Typescript file template based off of C# model